improve navbar and upload and download functily

+++ client/src/pages/Home.jsx
@@ -3,7 +3,12 @@ import { useSelector, useDispatch } from "react-redux";
 import { toast } from "react-toastify";
 import { selectFileList } from "../store/file/file.selector";
 import { selectCurrentUser } from "../store/user/user.selecter";
-import { addFile, deleteFile, fileDownloadApi, getFilesByUser } from "../api/file.api";
+import {
+  addFile,
+  deleteFile,
+  fileDownloadApi,
+  getFilesByUser,
+} from "../api/file.api";
 import Navbar from "../componets/Navbar";
 import FileContainer from "../componets/FileContainer";
 import { baseUrl } from "../api/apiUrl.api";
@@ -25,7 +30,11 @@ const Home = () => {
   }, []);
 
   const deleteFileHandle = async (_id) => {
-    await deleteFile(_id, currentUser.accessToken, dispatch);
+    if (confirm("Are you sure you want Delete File")) {
+      await deleteFile(_id, currentUser.accessToken, dispatch);
+    } else {
+      return;
+    }
   };
 
   const handleFile = (e) => {
@@ -41,16 +50,24 @@ const Home = () => {
   };
 
   const downloadFile = async (fileData) => {
-    const code = prompt("Please enter code:");
-    if (code === null) {
-      return;
-    }
-    if (code == fileData.code) {
-   const newFileData =await fileDownloadApi(fileData);
-   if(newFileData) FileDownload(newFileData, fileData.filename);
-    } else {
-      toast.error("code doesnot match");
-    }
+    // const code = prompt("Please enter code:");
+    // if (code === null) {
+    //   return;
+    // }
+    //   if (code == fileData.code) {
+    //  const newFileData =await fileDownloadApi(fileData);
+    //  if(newFileData) FileDownload(newFileData, fileData.filename);
+    //   } else {
+    //     toast.error("code doesnot match");
+    //   }
+
+    //  const newFileData =await fileDownloadApi(fileData);
+    //  if(fileData) FileDownload(fileData, fileData.filename);
+    const anchorElement = document.createElement("a");
+    anchorElement.href = `data:${fileData.mimetype};base64,${fileData.base64}`;
+    anchorElement.download = `${fileData.filename}`;
+    anchorElement.click();
+    anchorElement.remove();
   };
 
   return (
@@ -58,11 +75,8 @@ const Home = () => {
       <Navbar />
       <div>
         <div className="flex flex-col justify-between my-3 mx-5">
-          <h1 className="text-4xl font-bold italic text-center my-5">
-            File Keeper
-          </h1>
           <form className="flex items-center">
-            <input type="file" name="file" onChange={handleFile} />
+            <input type="file" name="file" onChange={handleFile} multiple />
             <button
               onClick={addFileHandler}
               className="border border-black rounded-lg p-2"
